refactor: move the api date helper up to the rides page
